TINY-7650: Update McAgar to not import TinyMCE and instead lazy load it
