+++ modules/mcagar/CHANGELOG.md
@@ -12,3 +12,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Upgraded to Katamari 8.0, which is incompatible with Katamari 7.0 if used in the same bundle.
+- TinyMCE is no longer imported and is instead loaded from `node_modules` if a `tinymce` global isn't available.
