+++ modules/mcagar/src/main/ts/ephox/mcagar/api/pipeline/TinyLoader.ts
@@ -2,7 +2,6 @@ import { TestLogs } from '@ephox/agar';
 import { Optional } from '@ephox/katamari';
 import { Insert, Remove, SugarBody, SugarElement, SugarShadowDom } from '@ephox/sugar';
 
-import 'tinymce';
 import * as Loader from '../../loader/Loader';
 import { setupTinymceBaseUrl } from '../../loader/Urls';
