+++ modules/mcagar/src/test/ts/browser/api/TinyLoaderTest.ts
@@ -45,7 +45,9 @@ UnitTest.asynctest('TinyLoader.setupInBodyAndShadowRoot passes logs through', (s
   TinyLoader.setupInBodyAndShadowRoot((_editor, onSuccess, _onFailure) => {
     calls++;
     onSuccess('call' + calls, TestLogs.single('log' + calls));
-  }, {}, (v, logs) => {
+  }, {
+    base_url: '/project/tinymce/js/tinymce'
+  }, (v, logs) => {
     try {
       if (SugarShadowDom.isSupported()) {
         Assert.eq('Value should come from second call', 'call2', v);
